requirements.txt
# Core dependencies
requests==2.31.0
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1

# NLP and ML
nltk==3.8.1
spacy==3.7.4
transformers==4.38.2

# Visualization
matplotlib==3.8.3

# Testing
pytest==8.0.2
faker==24.2.0

# Utilities
tqdm==4.67.1
python-dotenv==1.0.0  # For configuration management
pyyaml==6.0.1  # For YAML config files 